requirements.txt
Flask
requests
python-dotenv
opencv-python
Pillow
